Refactor file imports in ResultBox component and its related components (#7)
